tronWallet/transactions_demon/requirements.txt
psycopg2==2.9.2
flask_restful==0.3.9
python-dotenv==0.19.1
aio-pika==6.8.0
aiofiles==0.7.0
asyncpg==0.25.0
aiohttp==3.8.1
tronpy==0.2.6
pika==1.2.0








